clean up js for sorting method, add slight css
Sorting method now uses a simple sortBy(method) & renders so that
it can be used as standalone on the Library module. Normally it is
always called based on the dropdown input.

Still need to add localStorage as the next feauture (most important
functionability-wise). Also play around with colors and all css
(for background try colors like pastel(?) or beige & prob play w/
a color picker)